refactor(visual): split the palette from the conversation that fills it
The sixteen slots, the OSC 4 reply parser and the resolution a surface falls
back on are all arithmetic; only asking the terminal needs a terminal. The type
moves to rav-appearance and `query` stays here as a free function, so a target
that compiles its theme in can hold a Palette in read-only memory and never
write a byte to anything.

Moving it, I rewrote parse_rgb from its shape rather than carrying it across,
and quietly dropped three behaviours: trimming, the four-digit bound, and
refusing a fourth channel. A test caught the last one - `rgb:0/0/0/0` parsed as
black instead of being declined. The original is restored verbatim; the comment
explaining why guessing three of four channels is worse than declining is the
kind of thing a rewrite loses.
