+++ crates/rav-appearance/src/lib.rs
@@ -38,9 +38,11 @@ extern crate alloc;
 extern crate std;
 
 pub mod ink;
+pub mod palette;
 pub mod ramp;
 pub mod scene;
 
 pub use ink::{ANSI_NAMES, Colour, DEFAULT_ANSI, Ink};
+pub use palette::Palette;
 pub use ramp::{Ramp, Stripe};
 pub use scene::{Band, CapStyle, Scene, Style, StyleId};
