+++ src/visual/mod.rs
@@ -3,5 +3,7 @@
 pub mod palette;
 pub mod theme;
 
-pub use palette::Palette;
+// The palette itself is portable and lives in rav-appearance; only the OSC 4
+// conversation that fills one in needs a terminal, and that stays here.
+pub use rav_appearance::Palette;
 pub use theme::Theme;
